+++ app/api/game_service.py
@@ -67,11 +67,14 @@ class GameService:
 
         def team_state(team):
             scores = {}
+            timeouts_by_set = {}
             for i in range(1, session.sets_limit + 1):
                 scores[f"set_{i}"] = state.get_game(team, i)
+                timeouts_by_set[f"set_{i}"] = state.get_timeout(team, set_num=i)
             return TeamState(
                 sets=state.get_sets(team),
                 timeouts=state.get_timeout(team),
+                timeouts_by_set=timeouts_by_set,
                 scores=scores,
                 serving=(serve == State.SERVE_1 if team == 1 else serve == State.SERVE_2),
             )